backend_api/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
alembic
python-dotenv
pydantic
pydantic[email]
pydantic-settings